Use deleted_at when no delete action is found
We found some instances of deleted VMs that didn't have a delete
action recorded, despite having a deleted_at timestamp.

Since the invoicing code looks for a delete action to stop
charging for an instance, it would keep charging for deleted
instances that didn't have such an action.

Now it falls back to the deleted_at timestamp if no such an action
is found.

Related #21