Extend Instance SQL statement to fetch PCI and CPU
The information related to which GPU is in use by an instance
is available in the `pci_requests` column of the `instance_extra`
table.

For example:

```json
[
  {
    "count": 1,
    "spec": [
      {
        "product_id": "<REDACTED>",
        "vendor_id": "<REDACTED>",
        "dev_type": "type-PCI"
      }
    ],
    "alias_name": "V100",
    "is_new": false,
    "numa_policy": "legacy",
    "request_id": null,
    "requester_id": null
  }
]
```

The Instances table also contains `memory_mb` and `vcpus`.

It is therefore unnecessary to fetch any flavor information and
invoicing can be performed directly on the instance information.