Use the PCI and CPU/Mem info from the Instance
Instead of loading Flavor information from the flavors table
or from the flavors cache, we now use the information from
Instances and Instance Extra.

- `pci_info[0]["alias_name"]` from `instance_extra.pci_requests`
  contains the GPU name if the instance has one.
- `pci_info[0]["count"]` contains the GPU count.
- instances.memory_mb contains the RAM in MB.
- instances.vcpus contains the number of CVPUS.

If no PCI information the instance is assumed as CPU SU.

All flavors are 20GB, so that is hardcoded for the moment,
and root disk storage is not to be taken into account for
calculations of invoicing.

---

I have compared the values of January's billing values before
and after the change and they are the same.