Create k8s base dir with cronjob
Creates a cronjob that runs on the 1, 15th and 28th (mostly every
other week.)

The 1st is the most important one as that is what we'll be giving
out. This will generate the final invoice for the previous month.

15th will generate a mid-month usage invoice. Potentially useful
for notifying users about they currently have running.

28th will make sure that things still work a few days before
we need to bill on the 1st.