Allow ignoring shelved periods when billing VMs
When a VM is in a shelved state, we don't want to bill users.
`get_runtime_during()` has been refactored and corresponding test cases added.
Time spent in various VM states is now calculated by simulating the VM
as a finite state machine. Each state is represented as a `State` object
instance, which is entered upon certain trigger events (i.e the `Running`
state is entered on `unshelve`, `start`, or `create`).

We assume all state transitions are valid, and will not include
any validations for unusual event sequences (i.e `stopped` -> `unshelve`)

When in the `Error` state, the VM is not billable. New test cases,
mostly written by Kristi, have been added to reflect this desired behavior.
This behavior has been clarified in discussions and led to the discovery
of a bug regarding billing in the `Error` state [1].

[1] https://github.com/CCI-MOC/openstack-billing-from-db/pull/114#discussion_r2176195891