Fixed bug with stopped time between `stop` and `delete` event
When a VM is stopped, and afterwards deleted, the period of time
between the stop and delete event is not being added to the total
stopped time. The impact of this is undercounting of stopped time
for VMs which were stopped before they were deleted.

A test case is added, and `get_runtime_during()` has been fixed
and slightly refactored to be more readable